requirements.txt
ultralytics
opencv-python
numpy
fastapi
uvicorn
boto3
python-dotenv